updated to match functionallity of master
