+++ Lab2/out2
@@ -1,0 +1,9 @@
+000000 01001 01010 01000 00000 100000
+000000 01100 01101 01011 00000 100000
+000000 01111 11000 01110 00000 100000
+000000 10111 10110 11001 00000 100010
+000000 10101 10100 10101 00000 100010
+000000 10010 10001 10011 00000 100010
+000101 11101 11111 0000000000000000
+000000 11111 00101 11111 00000 100000
+000100 00100 00101 1111111111111110
