+++ Lab2/out3
@@ -1,0 +1,7 @@
+000000 00101 00110 00100 00000 100101
+000000 10000 01001 00100 00000 100100
+000000 10001 10010 10000 00000 100101
+000000 00000 00000 10000 00000 100100
+000010 00000000000000000000000000
+000101 00100 00101 1111111111111010
+000100 00110 00111 1111111111111001
