+++ Lab2/out4
@@ -1,0 +1,5 @@
+000000 00101 00110 00100 00000 100101
+000000 10000 01001 00100 00000 100100
+000000 10001 10010 10000 00000 100101
+000000 00000 00000 10000 00000 100100
+invalid instruction: xor
